pyproject.toml
[tool.poetry]
name = "python-sql"
version = "0.1.0"
description = ""
authors = ["polovykh <polovykh@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
